refactor: extract remaining UnifiedItemView components
- Extract UnifiedItemNutritionalInfo to separate component file
- Extract UnifiedItemFavorite to separate component file
- Clean up main UnifiedItemView file by removing duplicated components
- Add re-exports for backward compatibility
- Complete the componentization process for better maintainability
